Extracted preferences for holding tags into their own file.
