+++ src/Mod/Path/CMakeLists.txt
@@ -29,6 +29,7 @@ SET(PathScripts_SRCS
     PathScripts/PathDressupDragknife.py
     PathScripts/PathDressupHoldingTags.py
     PathScripts/PathDressupRampEntry.py
+    PathScripts/PathDressupTagPreferences.py
     PathScripts/PathDrilling.py
     PathScripts/PathEngrave.py
     PathScripts/PathFacePocket.py
